# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DenTeeth API Documentation",
        "description": "API Documentation for DenTeeth",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Development server"
        },
        {
            "url": "https://www.example.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get all appointments",
                "description": "Returns all appointments",
                "operationId": "09fd012ab97514481bde57bbeac2699b",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Create a new appointment",
                "operationId": "15017f04a457b26e2a870c8a011a407a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "service_id",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment added successfully"
                    },
                    "422": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get a specific appointment",
                "description": "Returns a specific appointment by ID",
                "operationId": "13b46cd03a1dae29a4c4dd6e756d37da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an existing appointment",
                "description": "Update an existing appointment",
                "operationId": "a779a3a6175fee8a68b75461718a1ec3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment updated successfully"
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete an appointment",
                "description": "Delete an appointment by ID",
                "operationId": "cc5727c249d22f6719923c09625213e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment deleted successfully"
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            }
        },
        "/api/diagnoses": {
            "get": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Get all diagnoses",
                "description": "Returns all diagnoses",
                "operationId": "6b5ce2fc3e77e59722fee40ad93b943b",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Create a new diagnosis",
                "description": "Create a new diagnosis",
                "operationId": "0383883ed7af41cccc6f1194efe6c5e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "symptom_data_id",
                                    "photo"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "symptom_data_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "photo_url.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Diagnosis added successfully"
                    },
                    "422": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/diagnoses/{id}": {
            "get": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Get a specific diagnosis",
                "description": "Returns a specific diagnosis by ID",
                "operationId": "6e165aaa0859fe89d27500d4164bd291",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Diagnosis not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Update an existing diagnosis",
                "description": "Update an existing diagnosis",
                "operationId": "5bb8ac2be732a4eac5a2dd3c19e3faa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "symptom_data_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "updated_photo.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Diagnosis updated successfully"
                    },
                    "404": {
                        "description": "Diagnosis not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Delete a diagnosis",
                "description": "Delete a diagnosis by ID",
                "operationId": "8913ee429b5e4f68a6ebf70e0c2d816e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diagnosis deleted successfully"
                    },
                    "404": {
                        "description": "Diagnosis not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}